scripts/cross_reference/requirements.txt
pyyaml>=6.0.1
pyahocorasick>=2.0.0
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0
requests>=2.31.0
networkx>=3.1
matplotlib>=3.7.2
pydantic>=2.4.2
fastapi>=0.100.0
uvicorn[standard]>=0.22.0